fixing file structure
