--- SwapStop-michael/requirements.txt
@@ -1,7 +1,0 @@
-fastapi
-uvicorn
-sqlalchemy
-pydantic
-alembic
-passlib
-bcrypt
